+++ src/circulo.h
@@ -3,18 +3,18 @@
 
 typedef void* CIRCULO;
 
-CIRCULO criaCirculo (int i, float x, float y, float r, char* corb, char* corp);
+CIRCULO criaCirculo (int i, double x, double y, double r, char* corb, char* corp);
 
 
 int getId_C (CIRCULO c);
 
-float getX_C (CIRCULO c);
+double getX_C (CIRCULO c);
 
-float getY_C (CIRCULO c);
+double getY_C (CIRCULO c);
 
-float getR_C (CIRCULO c);
+double getR_C (CIRCULO c);
 
-float calcularArea_C (CIRCULO c);
+double calcularArea_C (CIRCULO c);
 
 char* getCorB_C (CIRCULO c);
 
@@ -23,11 +23,11 @@ char* getCorP_C (CIRCULO c);
 
 void setId_C (CIRCULO c, int i);
 
-void setX_C (CIRCULO c, float x);
+void setX_C (CIRCULO c, double x);
 
-void setY_C (CIRCULO c, float y);
+void setY_C (CIRCULO c, double y);
 
-void setR_C (CIRCULO c, float r);
+void setR_C (CIRCULO c, double r);
 
 void setCorB_C (CIRCULO c, char* cb);
