+++ src/linha.c
@@ -7,12 +7,12 @@
 
 typedef struct{
     int i;
-    float x1, y1;
-    float x2, y2;
+    double x1, y1;
+    double x2, y2;
     char* cor;
 } linha;
 
-LINHA criaLinha (int i, float x1, float y1, float x2, float y2, char* cor){
+LINHA criaLinha (int i, double x1, double y1, double x2, double y2, char* cor){
     linha *l = (linha*) malloc (sizeof(linha));
 
     if (l == NULL){
@@ -43,22 +43,22 @@ int getId_L (LINHA l){
     return (l1->i);
 }
 
-float getX1_L (LINHA l){
+double getX1_L (LINHA l){
     linha *l1 = ((linha*) l);
     return (l1->x1);
 }
 
-float getY1_L (LINHA l){
+double getY1_L (LINHA l){
     linha *l1 = ((linha*) l);
     return (l1->y1);
 }
 
-float getX2_L (LINHA l){
+double getX2_L (LINHA l){
     linha *l1 = ((linha*) l);
     return (l1->x2);
 }
 
-float getY2_L (LINHA l){
+double getY2_L (LINHA l){
     linha *l1 = ((linha*) l);
     return (l1->y2);
 }
@@ -68,11 +68,11 @@ char* getCor_L (LINHA l){
     return (l1->cor);
 }
 
-float calcularArea_L (LINHA l){
+double calcularArea_L (LINHA l){
     linha *l1 = ((linha*) l);
     
-    float a = (l1->x2) - (l1->x1);
-    float b = (l1->y2) - (l1->y1);
+    double a = (l1->x2) - (l1->x1);
+    double b = (l1->y2) - (l1->y1);
 
     return ( 2 * sqrt(pow(a, 2) + pow(b, 2)));
 }
@@ -83,22 +83,22 @@ void setId_L (LINHA l, int i){
     l1->i = i;
 }
 
-void setX1_L (LINHA l, float x1){
+void setX1_L (LINHA l, double x1){
     linha *l1 = ((linha*) l);
     l1->x1 = x1;
 }
 
-void setY1_L (LINHA l, float y1){
+void setY1_L (LINHA l, double y1){
     linha *l1 = ((linha*) l);
     l1->y1 = y1;
 }
 
-void setX2_L (LINHA l, float x2){
+void setX2_L (LINHA l, double x2){
     linha *l1 = ((linha*) l);
     l1->x2 = x2;
 }
 
-void setY2_L (LINHA l, float y2){
+void setY2_L (LINHA l, double y2){
     linha *l1 = ((linha*) l);
     l1->y2 = y2;
 }
