+++ src/retangulo.h
@@ -3,20 +3,20 @@
 
 typedef void* RETANGULO;
 
-RETANGULO criaRetangulo (int i, float x, float y, float w, float h, char* corb, char* corp);
+RETANGULO criaRetangulo (int i, double x, double y, double w, double h, char* corb, char* corp);
 
 
 int getId_R (RETANGULO r);
 
-float getX_R (RETANGULO r);
+double getX_R (RETANGULO r);
 
-float getY_R (RETANGULO r);
+double getY_R (RETANGULO r);
 
-float getW_R (RETANGULO r);
+double getW_R (RETANGULO r);
 
-float getH_R (RETANGULO r);
+double getH_R (RETANGULO r);
 
-float calcularArea_R (RETANGULO r);
+double calcularArea_R (RETANGULO r);
 
 char* getCorB_R (RETANGULO r);
 
@@ -25,13 +25,13 @@ char* getCorP_R (RETANGULO r);
 
 void setId_R (RETANGULO r, int i);
 
-void setX_R (RETANGULO r, float x);
+void setX_R (RETANGULO r, double x);
 
-void setY_R (RETANGULO r, float y);
+void setY_R (RETANGULO r, double y);
 
-void setW_R (RETANGULO r, float w);
+void setW_R (RETANGULO r, double w);
 
-void setH_R (RETANGULO r, float h);
+void setH_R (RETANGULO r, double h);
 
 void setCorB_R (RETANGULO r, char* cb);
